DataChangedException is written to flymine.log whenever something suspicious happens. OSWFlymineImpl doesn't bother with DELETE on store when it know it just generated the id of the object. DataTracking test and StoreDataTestCase don't break because of changed data while removing data.

+++ intermine/src/java/org/intermine/objectstore/intermine/ObjectStoreWriterFlyMineImpl.java
@@ -174,6 +174,7 @@ public class ObjectStoreWriterFlyMineImpl extends ObjectStoreFlyMineImpl
         // to be atomic, even though it uses multiple statements.
         Connection conn = null;
         boolean wasInTransaction = isInTransaction();
+        boolean doDeletes = true;
         if (!wasInTransaction) {
             beginTransaction();
         }
@@ -182,6 +183,7 @@ public class ObjectStoreWriterFlyMineImpl extends ObjectStoreFlyMineImpl
             // Make sure this object has an ID
             if (o.getId() == null) {
                 o.setId(getSerial());
+                doDeletes = false;
             }
 
             // Make sure all objects pointed to have IDs
@@ -226,10 +228,12 @@ public class ObjectStoreWriterFlyMineImpl extends ObjectStoreFlyMineImpl
             while (cldIter.hasNext()) {
                 ClassDescriptor cld = (ClassDescriptor) cldIter.next();
                 String tableName = DatabaseUtil.getTableName(cld);
-                s.addBatch("DELETE FROM " + tableName + " WHERE id = " + o.getId());
-                logAddBatch();
-                //System//.out.println(getModel().getName() + ": Batched SQL:  DELETE FROM "
-                //        + tableName + " WHERE id = " + o.getId());
+                if (doDeletes) {
+                    s.addBatch("DELETE FROM " + tableName + " WHERE id = " + o.getId());
+                    logAddBatch();
+                    //System//.out.println(getModel().getName() + ": Batched SQL:  DELETE FROM "
+                    //        + tableName + " WHERE id = " + o.getId());
+                }
                 StringBuffer sql = new StringBuffer("INSERT INTO ")
                     .append(tableName)
                     .append(" (OBJECT");
